backend/requirements.txt
asgiref==3.8.1
chowkidar-strawberry==0.3.2
dj-database-url==2.3.0
Django==5.1.6
django-choices-field==2.3.0
django-cors-headers==4.7.0
graphql-core==3.2.6
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.260.2
strawberry-graphql-django==0.56.0
typing_extensions==4.12.2
whitenoise==6.9.0
